+++ Changelog.md
@@ -287,3 +287,9 @@ Since v2.5:
 - Fixed `sanitize_path`, `sanitize_auto` and `sanitize_legacy` incorrectly removing hyphens '`-`' from identifiers.
 - Fixed custom trident displaying incorrectly on shelves.
 - Added assetGen presets: `item_model/spear`, `item_model/player_head`, `item_model/handheld_rod`
+## 5.4
+- Added hook: `trim_pattern`
+- Added `modelNamespace`, `modelPathes` and `modelList` options.
+- Special and fallback models are now represented by special variant IDs.
+- Asset generator's inputPath now defaults to `"^.*$"`
+- Added `crawl` command